backend_requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# LLM & AI frameworks
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langgraph==0.0.20
openai==1.6.1

# Vector databases
chromadb==0.4.18
pinecone-client==2.2.4

# Knowledge graph
neo4j==5.15.0
py2neo==2021.2.4

# Text processing & embeddings
sentence-transformers==2.2.2
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2

# Background tasks
celery[redis]==5.3.4

# Integrations
atlassian-python-api==3.41.0
office365-rest-python-client==2.5.3
requests==2.31.0

# Utilities
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
typing-extensions==4.8.0

# Development & testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
flake8==6.1.0

# Monitoring & logging
structlog==23.2.0